Moved default schemas to be embedded in file instead of copied.
